requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Migrate==3.1.0
Pillow==8.3.1
markdown==3.3.4
python-dotenv==0.19.0
gunicorn==20.1.0
Flask-Bcrypt==0.7.1
email-validator==1.1.3
matplotlib==3.5.1
pandas==1.3.3
pygal==2.4.0
PyMySQL==1.0.2
cryptography==3.4.7
#管理依赖